GEODE-3436: Restore refactoring of IndexCommands
* See initial commit GEODE-3262 (ed293e817e547fb5ecd399bf4ba10d694af51e0a)
